% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomeHue.R
\name{biomeHue}
\alias{biomeHue}
\title{Get a vector of colours for a microbiota barplot}
\usage{
biomeHue(taxa)
}
\arguments{
\item{taxa}{A character vector of bacterial names for which colours are desired.}
}
\value{
A character vector of colours in hexadecimal format.
}
\description{
This function returns a vector of colours in hexadecimal format for a microbiota barplot. The colours are based on a pre-defined colour palette stored in the 'hue' data frame.
}
\examples{
# Get the colours for the bacteria 'Bacteroides' and 'Faecalibacterium'
biomeHue(c('Bacteroides', 'Faecalibacterium'))

# Get the full colour palette
biomeHue_palette()

}
