% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomeHue.R
\name{biomeHue_palette}
\alias{biomeHue_palette}
\title{Get the full colour palette for microbiota barplots}
\usage{
biomeHue_palette()
}
\value{
A data frame with two columns: 'Taxon' (the bacterial name) and 'Colour' (the colour in hexadecimal format).
}
\description{
This function returns the full pre-defined colour palette for microbiota barplots. The colour palette is stored in the 'hue' data frame.
}
\examples{
# Get the full colour palette
biomeHue_palette()

}
